# Optimize MapOf integer hash function